pushing diffs of mac copy onto remote repo
Pulling changes from remote repo onto mac. These changes will remove the
plugins dir for zsh config and zcomp file which is a dynamic file that
will always change and is based on the system it is installed on. These
two should be ignored symlinking the dirs into the .config dir.